# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Internship Transcript System V2 API
  version: 0.2.0
  summary: Target contract for internship evaluation, correspondence, and transcript generation
  description: |
    Canonical target contract for V2. This specification is derived from observed legacy workflows
    but intentionally replaces legacy /api/v1 shapes, unauthenticated CRUD, mutable templates,
    and synchronous email/PDF behavior.

    Unresolved organization policies remain marked TBD in descriptions and acceptance criteria.
    No Production credential belongs in this file.
  contact:
    name: Internship Transcript System Owner
  license:
    name: Proprietary / TBD
    identifier: LicenseRef-Proprietary
servers:
  - url: https://{apiHost}/api/v2
    description: Production placeholder
    variables:
      apiHost:
        default: api.example.invalid
        description: Inject the approved Production API hostname
...
security:
  - cookieAuth: []
paths:
...
  /students:
    get:
      tags: [Students]
      operationId: listStudents
      summary: List Students within effective scope
      x-permissions: [students.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: schoolId
          in: query
          schema:
            $ref: '#/components/schemas/ObjectId'
        - name: programId
          in: query
          schema:
            $ref: '#/components/schemas/ObjectId'
        - name: studentId
          in: query
          schema:
            type: string
        - name: evaluationStatus
          in: query
          schema:
            type: string
            enum: [awaiting_evaluator, awaiting_response, submitted, email_error]
      responses:
        '200':
          description: Paginated Students
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentList'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Students]
      operationId: createStudent
      summary: Create a Student
      x-permissions: [students.manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '201':
          description: Student created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '409':
          $ref: '#/components/responses/Conflict'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /students/import-preview:
    post:
      tags: [Students]
      operationId: previewStudentImport
      summary: Validate a student workbook without writing student records
      x-permissions: [students.import]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Import preview batch created; valid rows are not committed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentImportPreview'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          description: Workbook exceeds the 5 MiB upload or row limit
        '422':
          $ref: '#/components/responses/ValidationError'
  /students/imports/{batchId}:
    parameters:
      - name: batchId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    get:
      tags: [Students]
      operationId: getStudentImportPreview
      summary: Read an unexpired import preview owned by the caller
      x-permissions: [students.import]
      responses:
        '200':
          description: Current row decisions and commit outcomes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentImportPreview'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /students/imports/{batchId}/commit:
    parameters:
      - name: batchId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    post:
      tags: [Students]
      operationId: commitStudentImport
      summary: Commit up to 100 explicitly confirmed rows atomically per row
      x-permissions: [students.import]
      parameters:
        - name: idempotency-key
          in: header
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentImportCommitInput'
      responses:
        '200':
          description: Committed rows; retry with the same key and payload is safe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentImportCommit'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /students/{studentId}:
    parameters:
      - $ref: '#/components/parameters/StudentId'
    get:
      tags: [Students]
      operationId: getStudent
      summary: Get a Student within effective scope
      x-permissions: [students.read]
      responses:
        '200':
          description: Student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Students]
      operationId: updateStudent
      summary: Update a Student within effective scope
      x-permissions: [students.manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentPatch'
      responses:
        '200':
          description: Student updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Students]
      operationId: archiveStudent
      summary: Archive a Student without deleting history
      x-permissions: [students.manage]
      responses:
        '204':
          description: Student archived
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /organizations:
    get:
      tags: [Internship]
      operationId: listOrganizations
      summary: List internship organizations
      x-permissions: [organizations.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Archived'
      responses:
        '200':
          description: Paginated Organizations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Internship]
      operationId: createOrganization
      summary: Create an internship Organization
      x-permissions: [organizations.manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationInput'
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '422':
          $ref: '#/components/responses/ValidationError'
  /evaluators:
    get:
      tags: [Internship]
      operationId: listEvaluators
      summary: List external evaluators/advisers
      x-permissions: [organizations.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: organizationId
          in: query
          schema:
            $ref: '#/components/schemas/ObjectId'
      responses:
        '200':
          description: Paginated Evaluators
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluatorList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Internship]
      operationId: createEvaluator
      summary: Create an external Evaluator
      x-permissions: [organizations.manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluatorInput'
      responses:
        '201':
          description: Evaluator created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evaluator'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /placements:
    get:
      tags: [Internship]
      operationId: listPlacements
      summary: List Placements within effective scope
      x-permissions: [placements.read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: studentId
          in: query
          schema:
            $ref: '#/components/schemas/ObjectId'
        - name: termId
          in: query
          schema:
            $ref: '#/components/schemas/ObjectId'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PlacementStatus'
      responses:
        '200':
          description: Paginated Placements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlacementList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Internship]
      operationId: createPlacement
      summary: Create a Placement
      x-permissions: [placements.manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlacementInput'
      responses:
        '201':
          description: Placement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Placement'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: its_session
      description: Secure httpOnly same-site internal session cookie
    evaluatorBearer:
      type: http
      scheme: bearer
      bearerFormat: scoped-invitation-session
      description: Short-lived evaluator session bound to one assignment
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Search:
      name: search
      in: query
      schema:
        type: string
        minLength: 1
        maxLength: 200
    Archived:
      name: archived
      in: query
      schema:
        type: boolean
        default: false
...
    StudentId:
      name: studentId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthenticated:
      description: Authentication is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: AUTHENTICATION_REQUIRED
            message: Authentication is required.
            requestId: req_example
...
    Forbidden:
      description: Authenticated caller lacks action or scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: PERMISSION_DENIED
            message: You do not have permission to perform this action.
            requestId: req_example
    NotFound:
      description: Resource does not exist or is concealed from caller
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: RESOURCE_NOT_FOUND
            message: Resource not found.
            requestId: req_example
    Conflict:
      description: Duplicate, stale version, idempotency conflict, or invalid state transition
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request or publish validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: One or more fields are invalid.
            details:
              - field: email
                reason: invalid_format
            requestId: req_example
  schemas:
    ObjectId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
      example: 507f1f77bcf86cd799439011
...
    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
        requestId:
          type: string
    ErrorDetail:
      type: object
      required: [reason]
      properties:
        field:
          type: string
        reason:
          type: string
        message:
          type: string
    PageMeta:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    LocalizedText:
      type: object
      required: [th, en]
      properties:
        th:
          type: string
          maxLength: 500
        en:
          type: string
          maxLength: 500
...
    StudentInput:
      type: object
      required: [studentId, name, email, schoolId, programId]
      properties:
        studentId:
          type: string
          minLength: 3
          maxLength: 30
        name:
          $ref: '#/components/schemas/LocalizedText'
        email:
          type: string
          format: email
        personalEmail:
          type: string
          format: email
        schoolId:
          type: string
          minLength: 1
        programId:
          type: string
          minLength: 1
        courseId:
          type: string
        academicTermId: { type: string }
        semester: { type: string }
        company: { type: string }
        companyAddress: { type: string }
        province: { type: string }
        evaluatorName: { type: string }
        evaluatorEmail: { type: string, format: email }
        admissionYear: { type: integer, minimum: 2000 }
        academicYear: { type: integer, minimum: 2000 }
        status: { type: string, enum: [active, archived], default: active }
    StudentPatch:
      type: object
      properties:
        studentId:
          type: string
          minLength: 3
          maxLength: 30
        name:
          $ref: '#/components/schemas/LocalizedText'
        email:
          type: string
          format: email
        personalEmail:
          type: string
          format: email
        schoolId:
          type: string
          minLength: 1
        programId:
          type: string
          minLength: 1
        courseId:
          type: string
        academicTermId: { type: string }
        semester: { type: string }
        company: { type: string }
        companyAddress: { type: string }
        province: { type: string }
        evaluatorName: { type: string }
        evaluatorEmail: { type: string, format: email }
        admissionYear: { type: integer, minimum: 2000 }
        academicYear: { type: integer, minimum: 2000 }
        status: { type: string, enum: [active, archived] }
    Student:
      allOf:
        - $ref: '#/components/schemas/StudentInput'
        - type: object
          required: [id]
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            evaluationStatus:
              type: string
              enum: [awaiting_evaluator, awaiting_response, submitted, email_error]
            archivedAt:
              type: [string, 'null']
              format: date-time
    StudentList:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Student'
        meta:
          $ref: '#/components/schemas/PageMeta'
    StudentImportIssue:
      type: object
      required: [code, field, message]
      properties:
        code: { type: string }
        field: { type: string }
        message: { type: string }
    StudentImportChange:
      type: object
      required: [field, before, after]
      properties:
        field: { type: string }
        before: {}
        after: {}
    StudentImportItem:
      type: object
      required: [id, sheet, rowNumber, action, status, issues, warnings, changes]
      properties:
        id:
          $ref: '#/components/schemas/ObjectId'
        sheet: { type: string }
        rowNumber: { type: integer, minimum: 1 }
        student:
          type: object
          additionalProperties: true
        action:
          type: string
          enum: [create, update, unchanged, invalid]
        status:
          type: string
          enum: [pending, committed, skipped]
        outcome:
          type: string
          enum: [created, updated, unchanged, skipped]
        issues:
          type: array
          items: { $ref: '#/components/schemas/StudentImportIssue' }
        warnings:
          type: array
          items: { $ref: '#/components/schemas/StudentImportIssue' }
        changes:
          type: array
          items: { $ref: '#/components/schemas/StudentImportChange' }
    StudentImportSummary:
      type: object
      required: [total, creates, updates, unchanged, invalid, warnings, committed]
      properties:
        total: { type: integer, minimum: 0 }
        creates: { type: integer, minimum: 0 }
        updates: { type: integer, minimum: 0 }
        unchanged: { type: integer, minimum: 0 }
        invalid: { type: integer, minimum: 0 }
        warnings: { type: integer, minimum: 0 }
        committed: { type: integer, minimum: 0 }
    StudentImportPreview:
      type: object
      required: [batchId, sourceName, checksum, expiresAt, status, questionFieldsDetected, summary, items]
      properties:
        batchId:
          $ref: '#/components/schemas/ObjectId'
        sourceName: { type: string }
        checksum: { type: string, pattern: '^[a-f0-9]{64}$' }
        expiresAt: { type: string, format: date-time }
        status: { type: string, enum: [preview, committed] }
        questionFieldsDetected:
          type: array
          items: { type: string }
        summary:
          $ref: '#/components/schemas/StudentImportSummary'
        items:
          type: array
          maxItems: 1000
          items:
            $ref: '#/components/schemas/StudentImportItem'
    StudentImportCommitInput:
      type: object
      additionalProperties: false
      required: [decisions]
      properties:
        decisions:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            additionalProperties: false
            required: [rowId, action]
            properties:
              rowId:
                $ref: '#/components/schemas/ObjectId'
              action: { type: string, enum: [create, update] }
    StudentImportCommit:
      type: object
      required: [batchId, commitId, status, results]
      properties:
        batchId:
          $ref: '#/components/schemas/ObjectId'
        commitId:
          $ref: '#/components/schemas/ObjectId'
        status: { type: string, enum: [completed] }
        results:
          type: array
          items:
            type: object
            required: [rowId, action, outcome]
            properties:
              rowId:
                $ref: '#/components/schemas/ObjectId'
              action: { type: string, enum: [create, update] }
              outcome: { type: string, enum: [created, updated, unchanged] }
...
    OrganizationInput:
      type: object
      required: [name, address, province]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 300
        address:
          type: string
          maxLength: 1000
        province:
          type: string
          maxLength: 100
    Organization:
      allOf:
        - $ref: '#/components/schemas/OrganizationInput'
        - type: object
          required: [id, createdAt, updatedAt, version]
          properties:
            id:
              $ref: '#/components/schemas/ObjectId'
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
            version: { type: integer, minimum: 1 }
            archivedAt: { type: [string, 'null'], format: date-time }
    OrganizationList:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        meta:
          $ref: '#/components/schemas/PageMeta'
    EvaluatorInput:
      type: object
      required: [organizationId, name, email]
      properties:
        organizationId:
          $ref: '#/components/schemas/ObjectId'
        name:
          type: string
          minLength: 1
          maxLength: 300
        email:
          type: string
          format: email
        position:
          type: string
          maxLength: 200
    Evaluator:
      allOf:
        - $ref: '#/components/schemas/EvaluatorInput'
        - type: object
          required: [id, createdAt, updatedAt, version]
          properties:
            id:
              $ref: '#/components/schemas/ObjectId'
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
            version: { type: integer, minimum: 1 }
            archivedAt: { type: [string, 'null'], format: date-time }
    EvaluatorList:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Evaluator'
        meta:
          $ref: '#/components/schemas/PageMeta'
    PlacementStatus:
      type: string
      enum: [draft, active, completed, cancelled, archived]
    PlacementInput:
      type: object
      required: [studentId, organizationId, termId, startsAt, endsAt]
      properties:
        studentId:
          $ref: '#/components/schemas/ObjectId'
        organizationId:
          $ref: '#/components/schemas/ObjectId'
        termId:
          $ref: '#/components/schemas/ObjectId'
        startsAt:
          type: string
          format: date
        endsAt:
          type: string
          format: date
    Placement:
      allOf:
        - $ref: '#/components/schemas/PlacementInput'
        - type: object
          required: [id, status, createdAt, updatedAt, version]
          properties:
            id:
              $ref: '#/components/schemas/ObjectId'
            status:
              $ref: '#/components/schemas/PlacementStatus'
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
            version: { type: integer, minimum: 1 }
    PlacementList:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Placement'
        meta:
          $ref: '#/components/schemas/PageMeta'
...